using shared json spec
