+++ spec/codecs/json.rb
@@ -1,6 +1,7 @@
 require "logstash/codecs/json"
 require "logstash/event"
 require "insist"
+require_relative 'json_shared'
 
 describe LogStash::Codecs::JSON do
   subject do
@@ -18,41 +19,11 @@ describe LogStash::Codecs::JSON do
       end
     end
 
-    context 'time-specific' do
-      let :data do
-        { 'foo' => 'bar', '@timestamp' => '2014-03-08T12:34:56+0100', '@version' => 1, 'baz' => 'bax' }
-      end
-
-      it 'should decode' do
-        decoded = false
-        subject.decode(data.to_json) do |event|
-          decoded = true
-        end
-        decoded.should be_true
-      end
-
-      it 'should decode an event' do
-        subject.decode(data.to_json) do |event|
-          expect(event).to be_a LogStash::Event
-        end
-      end
-
-      it 'should return an event from data with v1-specific fields' do
-        subject.decode(data.to_json) do |event|
-          event['foo'].should eq(data['foo'])
-          event['baz'].should eq('bax')
-          event['@version'].should eq(1)
-        end
-      end
-
-      it 'should decode time to ::Time to UTC' do
-        subject.decode(data.to_json) do |event|
-          expect(event.timestamp).to be_a(::Time)
-          event.timestamp.should eq(LogStash::Time.parse_iso8601(data['@timestamp']))
-          event.timestamp.gmtime.hour.should eq(11) # alias #utc
-        end
-      end
+    let :data_suffix do
+      ""
     end
+    
+    include_context 'json_structure'
 
     it "should be fast", :if => ENV["SPEEDTEST"] do
       json = '{"message":"Hello world!","@timestamp":"2013-12-21T07:01:25.616Z","@version":"1","host":"Macintosh.local","sequence":1572456}'
